seminar2/requirements.txt
fastapi
uvicorn
opencv-python-headless
pywavelets
scipy
Pillow
numpy
matplotlib
